change lists

+++ HPPortal.Data/Models/Sale.cs
@@ -6,9 +6,9 @@ namespace HPPortal.Data.Models
     public partial class Sale
     {
         public int SaleId { get; set; }
-        public int M1 { get; set; }
-        public int M2 { get; set; }
-        public int M3 { get; set; }
+        public decimal M1 { get; set; }
+        public decimal M2 { get; set; }
+        public decimal M3 { get; set; }
         public int ProductId { get; set; }
         public int PartnerId { get; set; }
         public string QuarterYear { get; set; }
